Modifications for usage of memory mapped files
Initially this was meant to be a contained storage manager within
postgres. However, we discarded this option since the interface defined
in fd.c is not always accessed through the smgr layer, but accessed from
various parts of the code base.

Hence this implementation directly modifies existing functionality in
fd.c which makes it only usable when the persistent memory storage
backend is used. Therefore this is ment to stay as a separate branch.