Experimental changes for direct access to PM data
This changes implement pointer redirection from postgres shared and
local buffers to the actual mapped files in persistent memory (PM). This
is meant to be a proof of concept and not meant to work on all scenarios.
The way shared buffer modifications and buffer release is currently done
creates certain problems that are not easy to circumvent.